Created Test and Method for #1

+++ AdditionLogic/Addition.cs
@@ -4,7 +4,7 @@ namespace AdditionLogic
 {
     public static class Addition
     {
-        public static int Add(int firstNum, int secondNum)
+        public static double Add(int firstNum, int secondNum)
         {
             return firstNum + secondNum;
         }
